Add BatchProcessingTracker  and refactor batch processing logic
